Revert "Run docker_credentials_gcr in warmer (#1780)"
This reverts commit cf4822c31cbfd0a79b9a0c4a59dbc250d258b4c1.
